% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLmodel.r
\name{BuiltEpiVar}
\alias{BuiltEpiVar}
\title{Fonction to create interaction variable.}
\usage{
BuiltEpiVar(X, Y, method, listGenesSNP, nbcomp = 2)
}
\arguments{
\item{X}{a genotype matrix where columns are genetic markers and rows are samples.}

\item{Y}{phenotype values can be logical (1/0) or numericGLmodel.}

\item{method}{method to consider for interactions variable  construction :
\itemize{
\item{"GGEE":}{ Compute the gene-gene interaction variables for each gene pair by finding its Eigen-epistasis Component defined as the linear combination of Gene SNPs having the highest correlation with the phenotype.} 
\item{"PCA":}{ Compute the gene-gene interaction variables for each gene pair with Principal Component Analysis.}
\item{"PLS":}{ Compute the gene-gene interaction variables for each gene pair with Partial Least Square.}
\item{"CCA":}{ Compute the gene-gene interaction variables for a given gene pair with Canonical Correlation Analysis.}
}}

\item{listGenesSNP}{list containing the names of genetic marker for each group.}

\item{nbcomp}{number of components to consider to built interaction variables for "PCA", "PLS" and "CCA" methods (2 by default).}
}
\value{
Returns a list including:
\item{XBet}{a matrix where columns are interaction variables and rows are samples.}
\item{interLength}{a vector that indicate the number of interaction variable for each gene couple.}
}
\description{
The fonction create interaction variable from a genetic data set structured by group.
}

