% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLmodel.r
\name{plotGLmodel}
\alias{plotGLmodel}
\title{Plot GLmodel cross-validation results}
\usage{
plotGLmodel(x, zoom = FALSE, dotNb = 10)
}
\arguments{
\item{x}{an object inheriting from class "GLmodel" that contains cross-validation results.}

\item{zoom}{if zoom=TRUE, only a reduced number of lambda values to plot. The number depending of dotNb.}

\item{dotNb}{number of lambda values to plot if zoom=TRUE (10 by default)}
}
\description{
Plot results from the GLmodel cross-validation with indication of minimal and oneSE lamba values
}

