% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simGenoDataR}
\alias{simGenoDataR}
\title{Fonction to simulate genotype data from a real data set}
\usage{
simGenoDataR(data, listGenesData, GenesMainNb, GenesPairNb, GenesRemainNb,
  sizeGenesMain, sizeGenesPair, sizeGenesRemain, SameMainPair,
  AllMainPair = TRUE, sampling = "gene")
}
\arguments{
\item{data}{a matrix of real data with SNP in columns and individuals in rows.}

\item{sizeGenesMain}{a vector of integers specifying the sizes of the genes having main effects.}

\item{sizeGenesPair}{a vector of integers specifying the sizes of the genes having interaction effects. The size of the vector has to be an even number, pairs being defined with the gene taken two by two along the vector.}

\item{sizeGenesRemain}{a vector of integers specifying the sizes of the genes having none effects.}

\item{SameMainPair}{logical. If TRUE, genes with interation effects will also have main effects}

\item{AllMainPair}{logical. If FALSE, only the first gene of each interaction pair will have a main effect. SameMainPair has to be TRUE.}

\item{sampling}{character. Can either be "gene" or "size". If "gene" is chosen, the genes are randomly selected in the initial data set whereas if it's "size", the genes are selected in the data set depending on their size, defined randomly.}

\item{listGenes}{a list that indicate the names of the SNPs composing each gene.}
}
\value{
Returns a list including:
\item{X}{a matrix where columns are SNPs and rows are individuals.}
\item{listGenesSNP}{a list that indicate the names of the SNPs composing each gene.}
\item{MainEff}{a vector containing the names of genes having main effects.}
\item{GenePair}{a vector containing the names of genes having interaction effects.}
}
\description{
Fonction to simulate genotype data from a real data set
}

