% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simPheno}
\alias{simPheno}
\title{Fonction to simulate phenotype values}
\usage{
simPheno(X, listGenes, MainEff, GenePair, model = "SNPproduct",
  pvBeta = c(2, 2), pvGamma = c(2, 2), r2 = 0.2, causalSNPnb,
  causalSNPportion = NULL, beta0 = 0)
}
\arguments{
\item{X}{a matrix where columns are SNPs and rows are individuals.}

\item{listGenes}{a list that indicate the names of the SNPs composing each gene.}

\item{MainEff}{a vector containing the names of genes having main effects.}

\item{GenePair}{a vector containing the names of genes having interaction effects. The size of the vector GenePair has to be an even number, pairs being defined with the gene taken two by two along the vector.}

\item{model}{the model to consider to simulate interaction effects, either "SNPproduct" or "PCproduct".}

\item{pvBeta}{numerical vector of possible values for main effects regression coefficients.}

\item{pvGamma}{numerical vector of possible values for interaction effects regression coefficients.}

\item{r2}{numeric value between 0 and 1. Coefficient of determination.}

\item{causalSNPnb}{numerical value. number of SNP in each gene to considered as causal.}

\item{causalSNPportion}{value between 0 and 1. Portion of SNP in each gene to considered as causal (if causalSNPnb is NULL).}

\item{beta0}{numeric value; intercept coefficient. Default 0.}
}
\value{
Returns a list including:
\item{y}{vector of simulated phenotype continuous values}
\item{G}{matrix of the simulated main effects.}
\item{GG}{matrix of the simulated interaction effects.}
\item{R2T}{numerical value for the coefficient of determination R2 when considering the model containing simulated main and interaction effects.} 
\item{R2I}{numerical value for the coefficient of determination R2 when considering the model containing only simulated interaction effects.} 
\item{R2S}{numerical value for the coefficient of determination R2 when considering the model containing only simulated main effects.} 
\item{caract}{a list with the caracteristic elements of the simulation.}
}
\description{
The fonction simulate a quantitative phenotype assuming a gene-structure among the SNPs.
}

